fix(js): Address PR comments
