+++ packages/js/.eslintrc.cjs
@@ -14,6 +14,7 @@ module.exports = {
     ],
     'local-rules/no-class-without-style': 'error',
     'id-length': 'off',
+    '@typescript-eslint/no-shadow': 'off',
   },
   parserOptions: {
     project: './tsconfig.json',
