+++ packages/js/src/ui/components/Notification/index.ts
@@ -1,1 +1,1 @@
-export * from './NotificationList';
+export { NotificationList } from './NotificationList';
