+++ packages/js/src/ui/context/LocalizationContext.tsx
@@ -1,11 +1,8 @@
-import { ParentProps, createContext, createMemo, useContext } from 'solid-js';
+import { createContext, createMemo, ParentProps, useContext } from 'solid-js';
 import { defaultLocalization } from '../config/defaultLocalization';
 import { Path } from '../helpers/types';
 
-export type Localization = {
-  'inbox.title': string;
-  'notifications.emptyNotice': string;
-};
+export type Localization = Partial<Record<keyof typeof defaultLocalization, string>>;
 
 type LocalizationPath = Path<Localization>;
